+++ packages/config/src/chains.ts
@@ -74,6 +74,7 @@ import {
   usdc,
   usdcwh,
   usdt,
+  usdtksm,
   usdtwh,
   vastr,
   vbnc,
@@ -809,7 +810,7 @@ export const kusama = new Parachain({
 
 export const kusamaAssetHub = new Parachain({
   assets: [
-    ChainAsset.fromAsset(usdt, {
+    ChainAsset.fromAsset(usdtksm, {
       decimals: 6,
       ids: {
         id: 1984,
@@ -911,6 +912,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '170050401128744171791743427490841452054',
       },
+      symbol: 'xcAMPE',
     }),
     ChainAsset.fromAsset(dev, {
       address: '0x0000000000000000000000000000000000000802',
@@ -926,6 +928,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '65216491554813189869575508812319036608',
       },
+      symbol: 'xcLIT',
     }),
     ChainAsset.fromAsset(maos, {
       address: '0xffffffff06a14960bddc8610d0e694d39a3c682e',
@@ -934,6 +937,7 @@ export const moonbaseAlpha = new EvmParachain({
         id: '8812816049497534070792537960559175726',
         palletInstance: 48,
       },
+      symbol: 'xcMAOS',
     }),
     ChainAsset.fromAsset(otp, {
       address: '0xFfffffFfB3229c8E7657eABEA704d5e75246e544',
@@ -941,6 +945,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '238111524681612888331172110363070489924',
       },
+      symbol: 'xcOTP',
     }),
     ChainAsset.fromAsset(paring, {
       address: '0xFFFffFfF8283448b3cB519Ca4732F2ddDC6A6165',
@@ -948,6 +953,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '173481220575862801646329923366065693029',
       },
+      symbol: 'xcPARING',
     }),
     ChainAsset.fromAsset(pica, {
       address: '0xFFFfFFFF10DD5Fd142163a40Ce0dae8c56e2801f',
@@ -955,6 +961,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '22417088946346045371238623691600461855',
       },
+      symbol: 'xcPICA',
     }),
     ChainAsset.fromAsset(tt1, {
       address: '0xfFffFfFf75976211C786fe4d73d2477e222786Ac',
@@ -962,6 +969,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '156305701417244550631956600137082963628',
       },
+      symbol: 'xcTT1',
     }),
     ChainAsset.fromAsset(tur, {
       address: '0xfFffffFf6448d0746f2a66342B67ef9CAf89478E',
@@ -969,6 +977,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '133300872918374599700079037156071917454',
       },
+      symbol: 'xcTUR',
     }),
     ChainAsset.fromAsset(unit, {
       address: '0xFfFFfFff1FcaCBd218EDc0EbA20Fc2308C778080',
@@ -977,6 +986,7 @@ export const moonbaseAlpha = new EvmParachain({
         id: '42259045809535163221576417993425387648',
         palletInstance: 48,
       },
+      symbol: 'xcUNIT',
     }),
     ChainAsset.fromAsset(ftmwh, {
       address: '0x566c1cebc6A4AFa1C122E039C4BEBe77043148Ee',
@@ -984,6 +994,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         palletInstance: 48,
       },
+      symbol: 'WFTM',
     }),
     ChainAsset.fromAsset(hdx, {
       address: '0xFFFfFfff345Dc44DDAE98Df024Eb494321E73FcC',
@@ -991,6 +1002,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '69606720909260275826784788104880799692',
       },
+      symbol: 'xcHDX',
     }),
     ChainAsset.fromAsset(usdcwh, {
       address: '0xE5dE10C4b744bac6b783fAF8d9B9fDFF14Acc3c9',
@@ -998,6 +1010,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         palletInstance: 48,
       },
+      symbol: 'USDC',
     }),
     ChainAsset.fromAsset(agng, {
       address: '0xFFfFFFFF38794F8c53fC7325ea07463dc6965e20',
@@ -1005,6 +1018,7 @@ export const moonbaseAlpha = new EvmParachain({
       ids: {
         id: '75066649112131892397889252914026143264',
       },
+      symbol: 'xcAGNG',
     }),
   ],
   ecosystem: Ecosystem.AlphanetRelay,
@@ -1251,6 +1265,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '224821240862170613278369189818311486111',
       },
+      symbol: 'xcACA',
     }),
     ChainAsset.fromAsset(astr, {
       address: '0xFfFFFfffA893AD19e540E172C10d78D4d479B5Cf',
@@ -1258,6 +1273,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '224077081838586484055667086558292981199',
       },
+      symbol: 'xcASTR',
     }),
     ChainAsset.fromAsset(aseed, {
       address: '0xfFfFFFFF52C56A9257bB97f4B2b6F7B2D624ecda',
@@ -1279,6 +1295,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '165823357460190568952172802245839421906',
       },
+      symbol: 'xcBNC',
     }),
     ChainAsset.fromAsset(cfg, {
       address: '0xFFfFfFff44bD9D2FFEE20B25D1Cf9E78Edb6Eae3',
@@ -1286,6 +1303,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '91372035960551235635465443179559840483',
       },
+      symbol: 'xcCFG',
     }),
     ChainAsset.fromAsset(dai, {
       address: '0x06e605775296e851FF43b4dAa541Bb0984E9D6fD',
@@ -1300,6 +1318,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '124463719055550872076363892993240202694',
       },
+      symbol: 'xcDED',
     }),
     ChainAsset.fromAsset(dot, {
       address: '0xFfFFfFff1FcaCBd218EDc0EbA20Fc2308C778080',
@@ -1307,6 +1326,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '42259045809535163221576417993425387648',
       },
+      symbol: 'xcDOT',
     }),
     ChainAsset.fromAsset(eq, {
       address: '0xFffFFfFf8f6267e040D8a0638C576dfBa4F0F6D6',
@@ -1314,6 +1334,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '190590555344745888270686124937537713878',
       },
+      symbol: 'xcEQ',
     }),
     ChainAsset.fromAsset(eqd, {
       address: '0xFFffFfFF8cdA1707bAF23834d211B08726B1E499',
@@ -1321,6 +1342,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '187224307232923873519830480073807488153',
       },
+      symbol: 'xcEQD',
     }),
     ChainAsset.fromAsset(fil, {
       address: '0xfFFfFFFF6C57e17D210DF507c82807149fFd70B2',
@@ -1328,6 +1350,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '144012926827374458669278577633504620722',
       },
+      symbol: 'xcFIL',
     }),
     ChainAsset.fromAsset(glmr, {
       address: '0x0000000000000000000000000000000000000802',
@@ -1343,6 +1366,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '69606720909260275826784788104880799692',
       },
+      symbol: 'xcHDX',
     }),
     ChainAsset.fromAsset(ibtc, {
       address: '0xFFFFFfFf5AC1f9A51A93F5C527385edF7Fe98A52',
@@ -1350,6 +1374,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '120637696315203257380661607956669368914',
       },
+      symbol: 'xcIBTC',
     }),
     ChainAsset.fromAsset(intr, {
       address: '0xFffFFFFF4C1cbCd97597339702436d4F18a375Ab',
@@ -1357,6 +1382,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '101170542313601871197860408087030232491',
       },
+      symbol: 'xcINTR',
     }),
     ChainAsset.fromAsset(ldot, {
       address: '0xFFfFfFffA9cfFfa9834235Fe53f4733F1b8B28d4',
@@ -1364,6 +1390,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '225719522181998468294117309041779353812',
       },
+      symbol: 'xcLDOT',
     }),
     ChainAsset.fromAsset(manta, {
       address: '0xfFFffFFf7D3875460d4509eb8d0362c611B4E841',
@@ -1371,6 +1398,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '166446646689194205559791995948102903873',
       },
+      symbol: 'xcMANTA',
     }),
     ChainAsset.fromAsset(nodl, {
       address: '0xfffffffFe896ba7Cb118b9Fa571c6dC0a99dEfF1',
@@ -1378,6 +1406,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '309163521958167876851250718453738106865',
       },
+      symbol: 'xcNODL',
     }),
     ChainAsset.fromAsset(neuro, {
       address: '0xFfffffFfB3229c8E7657eABEA704d5e75246e544',
@@ -1385,6 +1414,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '238111524681612888331172110363070489924',
       },
+      symbol: 'xcNEURO',
     }),
     ChainAsset.fromAsset(peaq, {
       address: '0xFffFFFFFEC4908b74688a01374f789B48E9a3eab',
@@ -1392,6 +1422,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '314077021455772878282433861213184736939',
       },
+      symbol: 'xcPEAQ',
     }),
     ChainAsset.fromAsset(pha, {
       address: '0xFFFfFfFf63d24eCc8eB8a7b5D0803e900F7b6cED',
@@ -1399,6 +1430,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '132685552157663328694213725410064821485',
       },
+      symbol: 'xcPHA',
     }),
     ChainAsset.fromAsset(pen, {
       address: '0xffFFfFFf2257622F345E1ACDe0D4f46D7d1D77D0',
@@ -1406,6 +1438,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '45647473099451451833602657905356404688',
       },
+      symbol: 'xcPEN',
     }),
     ChainAsset.fromAsset(ring, {
       address: '0xFfffFfff5e90e365eDcA87fB4c8306Df1E91464f',
@@ -1413,6 +1446,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '125699734534028342599692732320197985871',
       },
+      symbol: 'xcRING',
     }),
     ChainAsset.fromAsset(sub, {
       address: '0xfFfFffFf43B4560Bc0C451a3386E082bff50aC90',
@@ -1426,7 +1460,9 @@ export const moonbeam = new EvmParachain({
       decimals: 6,
       ids: {
         id: '166377000701797186346254371275954761085', // registered XC20 id for ERC20 tokens
+        palletInstance: 110,
       },
+      symbol: 'xcUSDC',
     }),
     ChainAsset.fromAsset(usdcwh, {
       address: '0x931715FEE2d06333043d11F658C8CE934aC61D0c',
@@ -1434,6 +1470,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         palletInstance: 110,
       },
+      symbol: 'USDC',
     }),
     ChainAsset.fromAsset(usdtwh, {
       address: '0xc30E9cA94CF52f3Bf5692aaCF81353a27052c46f',
@@ -1447,7 +1484,9 @@ export const moonbeam = new EvmParachain({
       decimals: 6,
       ids: {
         id: '311091173110107856861649819128533077277',
+        palletInstance: 110,
       },
+      symbol: 'xcUSDT',
     }),
     ChainAsset.fromAsset(vastr, {
       address: '0xFffFffff55C732C47639231a4C4373245763d26E',
@@ -1455,6 +1494,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '114018676402354620972806895487280206446', // registered XC20 id for ERC20 tokens
       },
+      symbol: 'xcvASTR',
     }),
     ChainAsset.fromAsset(vdot, {
       address: '0xFFFfffFf15e1b7E3dF971DD813Bc394deB899aBf',
@@ -1462,6 +1502,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '29085784439601774464560083082574142143', // registered XC20 id for ERC20 tokens
       },
+      symbol: 'xcvDOT',
     }),
     ChainAsset.fromAsset(vfil, {
       address: '0xFffffFffCd0aD0EA6576B7b285295c85E94cf4c1',
@@ -1469,6 +1510,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '272547899416482196831721420898811311297', // registered XC20 id for ERC20 tokens
       },
+      symbol: 'xcvFIL',
     }),
     ChainAsset.fromAsset(vglmr, {
       address: '0xFfFfFFff99dABE1a8De0EA22bAa6FD48fdE96F6c',
@@ -1476,6 +1518,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '204507659831918931608354793288110796652', // registered XC20 id for ERC20 tokens
       },
+      symbol: 'xcvGLMR',
     }),
     ChainAsset.fromAsset(vmanta, {
       address: '0xFFfFFfFfdA2a05FB50e7ae99275F4341AEd43379',
@@ -1483,6 +1526,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '289989900872525819559124583375550296953', // registered XC20 id for ERC20 tokens
       },
+      symbol: 'xcvMANTA',
     }),
     ChainAsset.fromAsset(wbtc, {
       address: '0xE57eBd2d67B462E9926e04a8e33f01cD0D64346D',
@@ -1504,6 +1548,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '150874409661081770150564009349448205842',
       },
+      symbol: 'xcZTG',
     }),
     ChainAsset.fromAsset(pink, {
       address: '0xfFfFFfFf30478fAFBE935e466da114E14fB3563d',
@@ -1511,6 +1556,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '64174511183114006009298114091987195453',
       },
+      symbol: 'xcPINK',
     }),
     ChainAsset.fromAsset(stink, {
       address: '0xffffffff54c556bd1d0f64ec6c78f1b477525e56',
@@ -1518,6 +1564,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '112679793397406599376365943185137098326',
       },
+      symbol: 'xcSTINK',
     }),
     ChainAsset.fromAsset(apillon, {
       address: '0xffffffff8a9736b44ebf188972725bed67bf694e',
@@ -1525,6 +1572,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '184218609779515850660274730699350567246',
       },
+      symbol: 'xcNCTR',
     }),
     ChainAsset.fromAsset(wifd, {
       address: '0xffffffff2e1d1ac9ea1686255befe995b31abc96',
@@ -1532,6 +1580,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '61295607754960722617854661686514597014',
       },
+      symbol: 'xcWIFD',
     }),
     ChainAsset.fromAsset(wbtce, {
       address: '0xffffffff1b4bb1ac5749f73d866ffc91a3432c47',
@@ -1560,6 +1609,7 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '294342517635293430510451841925413680520',
       },
+      symbol: 'xcLAOS',
     }),
     ChainAsset.fromAsset(eurc, {
       address: '0xffffffffa608ec1332131289cbd8a97a0bb3d90f',
@@ -1567,6 +1617,15 @@ export const moonbeam = new EvmParachain({
       ids: {
         id: '220698173844663170714431919474506717455',
       },
+      symbol: 'xcEURC.s',
+    }),
+    ChainAsset.fromAsset(movr, {
+      address: '0xffffffff54c47f43713fd9211724362697f1a302',
+      decimals: 18,
+      ids: {
+        id: 112675423039561305557350799263187182338n,
+      },
+      symbol: 'MOVR.mr',
     }),
   ],
   ecosystem: Ecosystem.Polkadot,
@@ -1605,6 +1664,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '319623561105283008236062145480775032445',
       },
+      symbol: 'xcBNC',
     }),
     ChainAsset.fromAsset(crab, {
       address: '0xFFFffFfF8283448b3cB519Ca4732F2ddDC6A6165',
@@ -1612,6 +1672,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '173481220575862801646329923366065693029',
       },
+      symbol: 'xcCRAB',
     }),
     ChainAsset.fromAsset(csm, {
       address: '0xffFfFFFf519811215E05eFA24830Eebe9c43aCD7',
@@ -1619,6 +1680,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '108457044225666871745333730479173774551',
       },
+      symbol: 'xcCSM',
     }),
     ChainAsset.fromAsset(kar, {
       address: '0xFfFFFFfF08220AD2E6e157f26eD8bD22A336A0A5',
@@ -1626,6 +1688,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '10810581592933651521121702237638664357',
       },
+      symbol: 'xcKAR',
     }),
     ChainAsset.fromAsset(kbtc, {
       address: '0xFFFfFfFfF6E528AD57184579beeE00c5d5e646F0',
@@ -1633,6 +1696,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '328179947973504579459046439826496046832',
       },
+      symbol: 'xcKBTC',
     }),
     ChainAsset.fromAsset(kint, {
       address: '0xfffFFFFF83F4f317d3cbF6EC6250AeC3697b3fF2',
@@ -1640,6 +1704,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '175400718394635817552109270754364440562',
       },
+      symbol: 'xcKINT',
     }),
     ChainAsset.fromAsset(ksm, {
       address: '0xFfFFfFff1FcaCBd218EDc0EbA20Fc2308C778080',
@@ -1647,6 +1712,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '42259045809535163221576417993425387648',
       },
+      symbol: 'xcKSM',
     }),
     ChainAsset.fromAsset(lit, {
       address: '0xfffFFfFF31103d490325BB0a8E40eF62e2F614C0',
@@ -1654,6 +1720,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '65216491554813189869575508812319036608',
       },
+      symbol: 'xcLIT',
     }),
     ChainAsset.fromAsset(movr, {
       address: '0x0000000000000000000000000000000000000802',
@@ -1669,6 +1736,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '189307976387032586987344677431204943363',
       },
+      symbol: 'xcPHA',
     }),
     ChainAsset.fromAsset(pica, {
       address: '0xFffFfFFf7dD9B9C60ac83e49D7E3E1f7A1370aD2',
@@ -1676,6 +1744,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '167283995827706324502761431814209211090',
       },
+      symbol: 'xcPICA',
     }),
     ChainAsset.fromAsset(rmrk, {
       address: '0xffffffFF893264794d9d57E1E0E21E0042aF5A0A',
@@ -1683,6 +1752,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '182365888117048807484804376330534607370',
       },
+      symbol: 'xcRMRK',
     }),
     ChainAsset.fromAsset(sdn, {
       address: '0xFFFfffFF0Ca324C842330521525E7De111F38972',
@@ -1690,6 +1760,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '16797826370226091782818345603793389938',
       },
+      symbol: 'xcSDN',
     }),
     ChainAsset.fromAsset(teer, {
       address: '0xFfFfffFf4F0CD46769550E5938F6beE2F5d4ef1e',
@@ -1697,6 +1768,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '105075627293246237499203909093923548958',
       },
+      symbol: 'xcTEER',
     }),
     ChainAsset.fromAsset(tnkr, {
       address: '0xfFFfFffF683474B842852111cc31d470bD8f5081',
@@ -1704,6 +1776,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '138512078356357941985706694377215053953',
       },
+      symbol: 'xcTNKR',
     }),
     ChainAsset.fromAsset(tur, {
       address: '0xfFffffFf6448d0746f2a66342B67ef9CAf89478E',
@@ -1711,13 +1784,16 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '133300872918374599700079037156071917454',
       },
+      symbol: 'xcTUR',
     }),
-    ChainAsset.fromAsset(usdt, {
+    //  USDT from Kusama Ecosystem
+    ChainAsset.fromAsset(usdtksm, {
       address: '0xFFFFFFfFea09FB06d082fd1275CD48b191cbCD1d',
       decimals: 6,
       ids: {
         id: '311091173110107856861649819128533077277',
       },
+      symbol: 'xcUSDT',
     }),
     ChainAsset.fromAsset(xrt, {
       address: '0xFffFFffF51470Dca3dbe535bD2880a9CcDBc6Bd9',
@@ -1725,6 +1801,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '108036400430056508975016746969135344601',
       },
+      symbol: 'xcXRT',
     }),
     ChainAsset.fromAsset(vbnc, {
       address: '0xFFffffff3646A00f78caDf8883c5A2791BfCDdc4',
@@ -1732,6 +1809,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '72145018963825376852137222787619937732',
       },
+      symbol: 'xcvBNC',
     }),
     ChainAsset.fromAsset(vksm, {
       address: '0xFFffffFFC6DEec7Fc8B11A2C8ddE9a59F8c62EFe',
@@ -1739,6 +1817,7 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '264344629840762281112027368930249420542',
       },
+      symbol: 'xcvKSM',
     }),
     ChainAsset.fromAsset(vmovr, {
       address: '0xfFfffFfF98e37bF6a393504b5aDC5B53B4D0ba11',
@@ -1746,6 +1825,40 @@ export const moonriver = new EvmParachain({
       ids: {
         id: '203223821023327994093278529517083736593',
       },
+      symbol: 'xcvMOVR',
+    }),
+    ChainAsset.fromAsset(glmr, {
+      address: '0xffffffff1a49463978f19dfd6983f2fa1885c254',
+      decimals: 18,
+      ids: {
+        id: 34940389883188672429634411751708934740n,
+      },
+      symbol: 'GLMR.mb',
+    }),
+    ChainAsset.fromAsset(usdc, {
+      address: '0xffffffff2215880e56fe63a96e54e073757c3092',
+      decimals: 6,
+      ids: {
+        id: 45305549634539991528356533909445161106n,
+      },
+      symbol: 'xcUSDC.mb',
+    }),
+    ChainAsset.fromAsset(usdcwh, {
+      address: '0xffffffff9e27ab60fadfd33abf71b39a7445f7ed',
+      decimals: 6,
+      ids: {
+        id: 210223998887087232417477484708341610477n,
+      },
+      symbol: 'whUSDC.mb',
+    }),
+    // Bridged USDT from Moonbeam Ecosystem
+    ChainAsset.fromAsset(usdt, {
+      address: '0xffffffffc70260a3bf46a91cf87b6f5e6abba712',
+      decimals: 6,
+      ids: {
+        id: 264528715839624829046161791069655377682n,
+      },
+      symbol: 'xcUSDT.mb',
     }),
   ],
   ecosystem: Ecosystem.Kusama,