+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @moonbeam-network/xcm-sdk
 
+## 4.1.8
+
+### Patch Changes
+
+- Updated dependencies [[`f80b62d`](https://github.com/moonbeam-foundation/xcm-sdk/commit/f80b62decd8f2ce46a76a6eac7c08942abf4cde2)]:
+  - @moonbeam-network/xcm-config@4.4.3
+
+## 4.1.7
+
+### Patch Changes
+
+- Updated dependencies [[`1562cf9`](https://github.com/moonbeam-foundation/xcm-sdk/commit/1562cf904ef3ead551114d89887841a0645e6c7a)]:
+  - @moonbeam-network/xcm-config@4.4.2
+
+## 4.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`f615192`](https://github.com/moonbeam-foundation/xcm-sdk/commit/f6151927a1645ec3e9b08a786a68571631a25631)]:
+  - @moonbeam-network/xcm-config@4.4.1
+
+## 4.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`0cc2a1d`](https://github.com/moonbeam-foundation/xcm-sdk/commit/0cc2a1d1d253bb7374b56de1798a243408caa369)]:
+  - @moonbeam-network/xcm-config@4.4.0
+
 ## 4.1.4
 
 ### Patch Changes